Delay tint2 startup if the system is in Live mode.
In some machines and expecially in Live mode, the panel is started before the Window Manager is properly
realized, thus making the latter ignore some set properties that make the panel window sticky on every desktop.

This has been workarounded in 67bd464 but unfortunately the fix doesn't work as it should.
We are then delaying the panel startup of three seconds in Live systems.

This issue still needs a proper fix though.

Signed-off-by: Eugenio Paolantonio (g7) <me@medesimo.eu>